fix: support alias roles
