+++ test/wpt-jsdom/to-run.yaml
@@ -175,6 +175,9 @@ name/shadowdom/basic.html:
 name/shadowdom/slot.html:
   [fail, upstream issue with dom-accessibility-api package]
 
+# ---
+# targets accessibility APIs whose mappings have not been implemented in this test suite yet,
+# DIR: core-aam
 ---
 DIR: dpub-aam
 
@@ -184,10 +187,8 @@ manual/*:
     targets accessibility APIs whose mappings have not been implemented in this test suite yet,
   ]
 role/roles.html:
-  "role: doc-pagefooter":
-    [fail, upstream issue with dom-accessibility-api package]
-  "role: doc-pageheader":
-    [fail, upstream issue with dom-accessibility-api package]
+  "role: doc-pagefooter": [fail, upstream issue with aria-query package]
+  "role: doc-pageheader": [fail, upstream issue with aria-query package]
 
 # ---
 # these are currently DOM query tests
@@ -231,10 +232,12 @@ roles-dynamic-switch.tentative.window.html:
 roles-generic.html:
   "el-li-orphaned": [fail, upstream issue with dom-accessibility-api package]
 roles.html:
-  "el-hgroup": [fail, upstream issue with dom-accessibility-api package]
-  "el-img": [fail, upstream issue with dom-accessibility-api package]
-  "el-s": [fail, upstream issue with dom-accessibility-api package]
-  "el-search": [fail, upstream issue with dom-accessibility-api package]
+  "el-hgroup":
+    [fail, upstream issue with aria-query and dom-accessibility-api packages]
+  "el-s":
+    [fail, upstream issue with aria-query and dom-accessibility-api packages]
+  "el-search":
+    [fail, upstream issue with aria-query and dom-accessibility-api packages]
 roles.tentative.html:
   "el-input-checkbox-switch":
     [fail, upstream issue with dom-accessibility-api package]
@@ -242,9 +245,24 @@ table-roles.html:
   "el-th-in-row": [fail, upstream issue with dom-accessibility-api package]
 
 ---
+DIR: svg-aam
+
+name/comp_host_language_label.html:
+  [fail, upstream issue with aria-query and dom-accessibility-api packages]
+role/role-img.tentative.html:
+  "Label: el-image (label from image>title element)":
+    [fail, upstream issue with dom-accessibility-api package]
+  "el-image (generic, explicitly empty image>title element)":
+    [fail, upstream issue with dom-accessibility-api package]
+role/roles.html:
+  "el-a[xlink:href]": [fail, upstream issue with dom-accessibility-api package]
+  "el-g": [fail, upstream issue with dom-accessibility-api package]
+  "el-image": [fail, upstream issue with dom-accessibility-api package]
+
+---
 DIR: wai-aria
 
-idlharness.window.html: [timeout, not a test file]
+idlharness.window.html: [timeout, out of scope]
 manual/*:
   [
     timeout,
@@ -267,12 +285,20 @@ role/grid-roles.html:
   "orphaned columnheader outside the context of row":
     [fail, upstream issue with dom-accessibility-api package]
 role/list-roles.html:
+  "orphan div with listitem role":
+    [fail, upstream issue with dom-accessibility-api package]
   "orphan p with listitem role":
     [fail, upstream issue with dom-accessibility-api package]
 role/listbox-roles.html:
   "orphaned option outside the context of listbox":
     [fail, upstream issue with dom-accessibility-api package]
 role/menu-roles.html:
+  "orphan div with menuitem role":
+    [fail, upstream issue with dom-accessibility-api package]
+  "orphan div with menuitemcheckbox role":
+    [fail, upstream issue with dom-accessibility-api package]
+  "orphan div with menuitemradio role":
+    [fail, upstream issue with dom-accessibility-api package]
   "orphaned menuitem outside the context of menu/menubar":
     [fail, upstream issue with dom-accessibility-api package]
   "orphaned menuitemradio outside the context of menu/menubar":
@@ -286,18 +312,24 @@ role/menu-roles.html:
   "orphan button with menuitemradio role":
     [fail, upstream issue with dom-accessibility-api package]
 role/roles.html:
-  "role: generic": [fail, unknown]
-  "role: suggestion": [fail, unknown]
+  "role: suggestion":
+    [
+      fail,
+      upstream issue with aria-query package - WAI-ARIA 1.3 not implemented,
+    ]
 role/synonym-roles.html:
-  "image role == computedrole image": [fail, unknown]
+  "image role == computedrole image":
+    [fail, upstream issue with aria-query and dom-accessibility-api packages]
   "synonym img role == computedrole image":
-    [fail, upstream issue with dom-accessibility-api package]
+    [fail, upstream issue with aria-query and dom-accessibility-api packages]
   "directory role == computedrole list":
-    [fail, upstream issue with dom-accessibility-api package]
+    [fail, upstream issue with aria-query and dom-accessibility-api packages]
   "div w/directory role == computedrole list":
-    [fail, upstream issue with dom-accessibility-api package]
+    [fail, upstream issue with aria-query and dom-accessibility-api packages]
 role/tab-roles.html:
   "orphan button with tab role":
+    [fail, upstream issue with dom-accessibility-api package]
+  "orphan span with tab role":
     [fail, upstream issue with dom-accessibility-api package]
 role/tree-roles.html:
   "orphaned treeitem outside the context of tree":
